backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
authlib
httpx
pydantic
pydantic-settings
python-dotenv
requests
psycopg2-binary
google-generativeai
